Conflicts: (solved)
	ini/trakem2/display/AreaList.java

+++ ini/trakem2/persistence/Loader.java
@@ -3920,6 +3920,7 @@ abstract public class Loader {
 		final String path = getAbsolutePath(p);
 		if (null == path) return null;
 		int i = path.length() -1;
+		// Safer than lastIndexOf: never returns -1
 		while (i > -1) {
 			if ('/' == path.charAt(i)) {
 				break;
@@ -4100,6 +4101,14 @@ abstract public class Loader {
 	/** Serializes the given object into the path. Returns false on failure. */
 	public boolean serialize(final Object ob, final String path) {
 		try {
+			// 1 - Check that the parent chain of folders exists, and attempt to create it when not:
+			File fdir = new File(path).getParentFile();
+			if (null == fdir) return false;
+			if (!fdir.exists() && !fdir.mkdirs()) {
+				Utils.log2("Could not create folder " + fdir.getAbsolutePath());
+				return false;
+			}
+			// 2 - Serialize the given object:
 			final ObjectOutputStream out = new ObjectOutputStream(new FileOutputStream(path));
 			out.writeObject(ob);
 			out.close();
@@ -4966,4 +4975,15 @@ abstract public class Loader {
 
 	/** Does nothing unless overriden. */
 	public void queueForMipmapRemoval(final Patch p, boolean yes) {}
+
+	/** Get the Universal Near-Unique Id for the project hosted by this loader. */
+	public String getUNUId() {
+		// FSLoader overrides this method
+		return Long.toString(System.currentTimeMillis());
+	}
+
+	// FSLoader overrides this method
+	public String getUNUIdFolder() {
+		return "trakem2." + getUNUId() + "/";
+	}
 }
